fix: добавление доп. проверки "if"; объединение модулей; стиль кода
